Move InnerNav controller into separate file
